grading-system/requirements.txt
Flask
tools
pymupdf
numpy
python-dotenv
sentence-transformers
google-genai
google-cloud-storage
google-cloud-aiplatform
google-cloud-vision